Move AddressBase and Onsud models into a generic uk_geo_utils app
The plan here is to move the generic tools for working with postcode/address
data out into a seperate django app. Once we've done that, we can pull that
app out of WhereDoIVote and package it as a component that can, in principle,
be shared by other projects. This is not as simple as renaming the addressbase
app because it contains a mix of generic code and code that is specific to
this project.

Moving the models is the first stage of that.
