+++ polling_stations/apps/data_finder/fixtures/test_addressbase.json
@@ -110,7 +110,7 @@
             "pfa": "",
             "imd": ""
         },
-        "model": "addressbase.onsud",
+        "model": "uk_geo_utils.onsud",
         "pk": "00000004"
     },
     {
@@ -140,7 +140,7 @@
             "pfa": "",
             "imd": ""
         },
-        "model": "addressbase.onsud",
+        "model": "uk_geo_utils.onsud",
         "pk": "00000005"
     },
 
@@ -171,7 +171,7 @@
             "pfa": "",
             "imd": ""
         },
-        "model": "addressbase.onsud",
+        "model": "uk_geo_utils.onsud",
         "pk": "00000007"
     },
     {
@@ -201,7 +201,7 @@
             "pfa": "",
             "imd": ""
         },
-        "model": "addressbase.onsud",
+        "model": "uk_geo_utils.onsud",
         "pk": "00000008"
     },
     {
@@ -231,7 +231,7 @@
             "pfa": "",
             "imd": ""
         },
-        "model": "addressbase.onsud",
+        "model": "uk_geo_utils.onsud",
         "pk": "00000009"
     }
 ]
